ARM: dts: mx6ul-symphony: don't use external watchdog pin
On the Symphony Board, you will not only need to connect two pins to get
the the external watchdog functionality working, R136 also needs to be
present.
R136 is not soldered on most (all?) Symphony Boards as it also would
have ramifications with some Ethernet functionality to add it.

Therefore change behaviour to use the internal watchdog for MX6UL* on
Symphony Board.

This change means that the `reboot` command will once again work
immediately, also on MX6UL Symphony boards.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>