arm64: dts: imx8mp-var-som: Add support for TSC2046 touchscreen
The VAR-SOM-MX8MP integrates a TSC2046 resistive touchscreen controller.
The controller is physically located on the SOM, and its signals are
routed to the SOM pins, allowing carrier boards to make use of it.

This patch adds the TSC2046 node and the appropriate SPI controller.

Upstream-Status: Backport [e583afbe3568c654a221793019e77738ffd2946f]

Signed-off-by: Stefano Radaelli <stefano.r@variscite.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>
Signed-off-by: Stefano Radaelli <stefano.r@variscite.com>